(slime-swank-connection-retries): Increased default number of connection retries from 5 to ten.
Indentation "cleanups": somehow I was using
`common-lisp-indent-function' for Emacs Lisp code previously.

(slime-edit-fdefinition): Added M-. (edit definition) and M-, (pop
definition stack) commands. Definitions are found in much the same way
Hemlock does it. The user-interface is not the same as TAGS, because I
like this one better. We can add TAGS-compatibility as an optional
feature in future.