Moved many CL fragments from slime.el to swank.lisp.
(slime-compile-file, slime-compile-defun, slime-goto-location):
Compiler notes are now represented with a property list.  To find the
source expression first move to the file offset of the top-level form
and then use the source path to find the expression.  This should
avoid many reader issues.  For compile-defun store the start position
of the top-level expression from the buffer in the compiler notes and
use that to locate error messages.  Add error overlays for notes
without context to the first available expression.

(slime-process-available-input): Don't protect from reader errors.