(briefly-describe-symbol-for-emacs): Add support for alien-types. (describe-alien-type, %describe-alien, describe-alien-struct,  describe-alien-union,  describe-alien-enum): New functions.
(source-path-file-position): Read the entire expression with special
readtable.  The readtable records source positions for each read
sub-expression in a hashtable.  Extract the subexpression for the
source path from the read object and lookup the subexpression in the
hashtable to find its source position.

(read-and-record-source-map,
make-source-recorder, make-source-recording-readtable,
make-source-map, *source-map*, lookup-recording-readtable,
source-path-stream-position, find-path-in-form,
find-form-in-source-map)  New functions.