(code-location-stream-position): Position the argument stream at the definition before returning.
(source-location-from-code-location): Include the :snippet hint for
Emacs (see above). The snippet will only be accurate provided that the
source file on disk has not been modified.

(*source-file-cache*) The contents of all source files consulted for
M-. are now cached if they match the version of the running code. This
is so that we can accurately lookup source locations even when the
file is modified, provided we manage to get the right version (by file
timestamp) at least once.

(source-location-from-code-location): If the right source version is
not available on disk or in our cache then let Emacs fall back on a
regular expression search.