(spawn-repl-thread): Use *default-worker-thread-bindings* just like spawn-worker-thread.
(wrap-sldb-vars): New function.  When evaluating a form in a frame,
Allegro uses the "old" dynamic context, i.e., dynamic variables like
*sldb-level* and the like are reset to the values in those frames.
But if *sldb-level* is reset to 0, Emacs doesn't notice when a
(recursive) error occurs while evaluating the form in the old frame.
wrap-sldb-vars saves the debugger related variables to avoid such
confusion.
(eval-string-in-frame, pprint-eval-string-in-frame): Use it.