Work on repl history navigation, restore old M-p/M-n behaviour due to #lisp demand
Also print the current regexp in the minibuffer messages.
M-p/M-n takes the repl input up to the point not the entire input as it did before.
slime-repl-previous/next-input-starting-with-current-input: new names for the old M-p/M-n commands
History navigation commands jump to the end of buffer when point is before the prompt.