Unbreak "copy-to-repl" functionality in slime-presentations.el
The `slime-copy-presentation-to-repl' function simply copied a
presentations text to the REPL, which made these objects
ephemeral. For *inspector* the presentation was mispointing as soon as
the user navigated away*. For *sldb* the presentation it just didn't
work.

For this functionality, slime-presentation.el uses the same strategy
as slime-repl.el. A new `sldb-copy-down-to-repl' is also offered in
slime-repl.el.

* contrib/slime-presentations.el (slime-copy-presentation-to-repl):
Unbreak. Use same `slime-repl-send-string' strategy as
slime-repl.el.

* contrib/slime-repl.el (sldb-mode-map): Bind M-RET to
`sldb-copy-down-to-repl' in `sldb-mode-map'.
(sldb-copy-down-to-repl): New function, much like
`slime-inspector-copy-down-to-repl'

* contrib/slime-repl.el (slime-inspector-copy-down-to-repl): Use only
first return value of SWANK:INSPECTOR-NTH-PART.