More adjustments to find-definition in ACL backend
Previous changes broke compilation for definitions in buffers that
don't have files(like *slime-scratch*). Finding definitions in buffers
that have a file name that doens't exist in the file-system was also
broken.

Also WHO-CALLS functionality for top-level forms is unreliable in ACL,
be it for temp or other loading mechanisms. For the former, note
*TEMP-FILE-MAP* might point to dead buffers and for the latter, there
is not enough file info in the .fasl annotations. See
http://franz.com/support/documentation/ \
current/doc/cross-reference.htm#in-functions-arg-2.

Therefore, this change also deletes the previous attempt by
FSPEC-DEFINITION-LOCATIONS to reliably follow :TOP-LEVEL-FORM
references..

* swank-allegro.lisp (fspec-definition-locations): Don't attempt
to interpret :TOP-LEVEL-FORM fspecs, since file info is always
unreliable.

* swank-allegro.lisp (compile-from-temp-file): Reinstate recording
in *TEMP-FILE-MAP* for all compilations. It is still useful if we
haven't saved the buffer or are operating from a file-less buffer,
like *slime-scratch*

(cherry picked from commit 6b3577823881826b42802db44c2197dbff9377b2)

Conflicts:
	ChangeLog