Closes #137: Inspecting * in REPL no longer uses second-to-last value
The previous technique for setting *, ** and friends just-in-time
didn't work, especially when the inspector was called from the REPL
buffer to inspect *. The problem this was fixing was fixed in a
different manner, by making contrib/slime-repl.el ensure that these
variables are set in the swank side.

* contrib/slime-repl.el (sldb-copy-down-to-repl): Call new
SWANK:SET-REPL-VARIABLES just after copying down to REPL.

* contrib/swank-repl.lisp (set-repl-variables): New slimefun.
(repl-eval): Call SET-REPL-VARIABLES *after* evaluating region.