Print a note everytime objects are returned to the REPL
Also some sly-mrepl groundwork for returning calls to REPL

* contrib/sly-mrepl.el (sly-mrepl--copy-to-repl): New function.
(sly-mrepl--eval-for-repl): Call `sly-mrepl--copy-to-repl'
(:copy-to-repl): Add `note' arg and insert it after the ";;"
(sly-mrepl--eval-for-repl): Add `note' arg. Send note to lisp.
(sldb-copy-down-to-repl, sly-trace-dialog-copy-down-to-repl)
(sly-inspector-copy-down-to-repl): Pass a note to
`sly-mrepl--eval-for-repl'
(sly-inspector-copy-down-to-repl): Use
SWANK:INSPECTOR-NTH-PART-OR-LOSE

* contrib/swank-mrepl.lisp (mrepl-get-history-entry): New function.
(mrepl-get-object-from-history): Use it.
(globally-save-object): Saves SLAVE-SLYFUN's multiple return
values. Pass a note to COPY-VALUES-TO-REPL
(:copy-to-repl): Can copy multiple values to REPL.
(copy-values-to-repl): Add NOTE arg.
(:sync-package-and-default-directory): Pass a note to
COPY-VALUES-TO-REPL.
(*saved-objects*, *saved-objects-note*): New variables

* lib/lisp/swank.lisp (inspector-nth-part-or-lose): New
slyfun. Wonder if the other one is still useful.