Get rid of the ugly `sly-mrepl--copy-to-repl-after' hack
* contrib/sly-mrepl.el (sly-mrepl--copy-to-repl-after): Removed
hackish variable.
(sly-mrepl-mode): Ditto.
(:copy-to-repl): Delete channel method.
(sly-mrepl--insert-prompt): `condition' arg is optional.
(sly-mrepl--eval-for-repl): New function (little relation to the
old function with the same name.)
(sly-mrepl--copy-objects-to-repl): Use `sly-mrepl--eval-for-mrepl'
(sly-mrepl--save-and-copy-for-repl): Renamed from
`sly-mrepl--eval-for-mrepl'.
(sly-inspector-copy-part-to-repl, sly-db-copy-part-to-repl)
(sly-apropos-copy-symbol-to-repl)
(sly-trace-dialog-copy-part-to-repl, sly-db-copy-call-to-repl)
(sly-trace-dialog-copy-call-to-repl, sly-mrepl-set-directory): Use
`sly-mrepl--save-and-copy-for-repl'.
(sly-mrepl-set-package): Use `sly-mrepl--eval-for-repl'.

* contrib/slynk-mrepl.lisp (copy-values-to-repl): Deleted.
(prompt-arguments): New helper.
(send-prompt): Use PROMPT-ARGUMENTS.
(:copy-to-repl): Removed channel method.
(:sync-package-and-default-directory): Removed channel method.
(guess-and-set-package): Moved slyfun down.
(eval-for-mrepl): Add docstring and clarify behaviour.
(copy-to-repl, sync-package-and-default-directory): New "slave" slyfuns.