Buffer compilation arms sly-stickers too
* sly.el (sly-region-for-defun-at-point): Accept POS as default point.

* contrib/sly-stickers.el
(sly-stickers-compile-region-aware-of-stickers-1): New helper
function. Accept `callback'. Accept `sync', `flash` and `fallback`
as boolean kwargs.
(sly-stickers-compile-region-aware-of-stickers): Make use of
`sly-stickers-compile-region-aware-of-stickers-1'
(sly-stickers-after-buffer-compilation): Implemented, makes use of
`sly-stickers-compile-region-aware-of-stickers-1'.

* contrib/slynk-stickers.lisp (defpackage :slynk-stickers): Export slyfuns
explicitly to appease sbcl.
(compile-for-stickers): Add docstring. Only return PROBE if it was
a successful compilation. Return nil if no ORIGINAL-STRING was
given.

* test/sly-stickers-tests.el (stickers-in-a-file): New test.
(stickers-record-stuff): Fix FIXME.