WIP #21: Fix some incremental output issues without extra thread
This is a tentative change for various reasons:

- In slynk/slynk-gray.lisp a similar technique to be use here (git
  8a91fe75 Nikodemus Siivola 2006-04-12 08:43:55 +0000 32) but was
  removed in 2008. Bringing it back seems harmless enough, but only in
  limited tests. Also STREAM-WRITE-STRING does not implement flushing
  on newline.

- Some Lisps completely ignore the BUFFERING arg to ACCEPT-CONNECTION.

- All this is trying to avoid a dedicated thread for flushing all the
  output streams, which, being a dirty hack, might in the end be a
  the most battle-tested alternative.

Also the default value for SWANK-MREPL:*USE-DEDICATED-OUTPUT-STREAM*
should probably be moved to a function that detects if connection is
between the same host or not. This is actually a separate problem,
but detected it while editing the manual.

* contrib/slynk-mrepl.lisp (*dedicated-output-stream-buffering*): Default
to :line, not t. SBCL and CMUCL seem to abide. CCL and ALLEGRO
ignore it.

* doc/sly.texi (REPL output control): New section.

* slynk/slynk-gray.lisp (stream-write-char): For the STREAM-WRITE-CHAR
interface, FINISH-OUTPUT when a newline is detected.

* doc/sly.texi: Remove sub menus that make it very hard to edit and
restructure the manual.