Fix stream inconsistency problem when restarting images
In REVERT-GLOBAL-IO-REDIRECTION and related function, must be sure to
not signal errors. In this case SLYNK:*LOG-OUTPUT* must be used
instead of *STANDARD-OUT*, since the standard streams whose
redirection are about to be reverted might be in an unconsistent state
after, for instance, restarting an image.

* slynk-mrepl.lisp (open-dedicated-output-stream): Improve doc slightlyt.
(revert-global-io-redirection): Log to SLYNK:*LOG-OUTPUT*.