Implement set-stream-timeout on sbcl.
This is primarily used when reading password at connect time. For
implementation that do not define that, an attacker can open connection
to the port and by this prevent other from access to the lisp process.

With this in place, attacker still can start sending absurdly large
chunk of data and send a single octet every 20 seconds, but it is at
least some protection. Better way might be handling passwords in the
separate thread, or using with-deadline, but this infrastructure is
already in place.