slime-version: check if we're opening slime.el
When byte compiling a contrib via the Makefile, slime.el is loaded (via
`require') but not compiled. That causes `byte-compile-current-file' to
point to the contrib file, rather than slime.el, generating confusion
when setting `slime-version'. To avoid this, let's look at both
`load-file-name' and `byte-compile-current-file' to see if one of them
points to slime.el. (Sadly, this involves hard-coding the file name.)