Make calling slime-setup unnecessary
For a while now, the party line has been that slime-contribs is the way
to select which contribs are to be loaded. However, there were two
common scenarios that didn't work properly:

  1. (require 'slime)
     (setq slime-contribs <something>)

  2. Installing SLIME via package.el, then setting slime-contribs.

In both cases, the problem was that, after loading SLIME, changing
slime-contribs has no effect since it was only being taken into account
at load-time. This commit makes it so that we now look at slime-contribs
each time a new SLIME connection is launched.