abcl: radical improvements to inspection and source locations
Most of the functionality in this patch requires the as yet unreleased
abcl-1.5.0 which may be built from the
[abcl-1.5.0-dev][abcl-source-svn] for which [gitlab.common-lisp.net
provides a convenient git
bridge](https://gitlab.common-lisp.net/abcl/abcl)

[abcl-source-svn]: svn://abcl.org/svn/trunk/abcl

Enable goto source location to find definitions in JAR archives.

For primitives compiled from Java source, return implementation source
location designated by the logical pathname associated with 'SYS:SRC'.

Finding functions and methods works well, other than when methods are
defined as part of a DEFGENERIC form, or implicitly as with :report on
condition defining PRINT-OBJECT.

FIND-DEFINITIONS include symbols and specials defined in Java.

Redirect source lookup to system jar if it's not where it claims to be.

Show stack trace when inspecting Java conditions.

Implement an inspector for Java classes.

Make line labels in inspector (:labels) so they display nicely.

Allow find definitions to be called even if the symbol isn't found in
the current package, if *find-definitions-all-packages* is t (as it
will be for ABCL). Default nil

Print internal variables in Java stack traces.

Prettier function name if we can glean information about otherwise
anonymous function.

Fix so that edit slot action in clos object inspector works.

Tweak to invoke hyperspec in inspector (offered in symbol browser if a
CL symbol) - *slime-inspector-hyperspec-in-browser* if t shows
hyperspec in emacs buffer despite setting of
browse-url-browser-function.

Don't use the ABCL native printObject for frames, so we can get prettier
function display.

Compile DEFIMPLEMENTATION forms so that they are visible to edit definition.

Fix source-location (used by view frame source) to work in general and
in particular for local function - go to the source of their owner.

Don't show bogus locals for Java stack frames.