slime-c-p-c-tests: add tests for slime-complete-symbol*
These tests verify that `slime-complete-symbol*' correctly inserts the
desired completion in the target buffer.

The tests currently fail for completion prefixes with more than one
candidate completion. For example, if one calls
`slime-complete-symbol*' with point just after the prefix "map" and
chooses "mapcar" as the desired completion from the *Completions*
buffer, then the resulting completed text will be "mapmapcar", not the
desired "mapcar".

This commit contains tests to demonstrate the bug. The fix will be
submitted in a follow-on commit.

This commit also renames the old `complete-symbol*' tests to
`completions', since they don't actually test
`slime-complete-symbol*', but rather `slime-completions'.