Fix loader for implementations which don't modify packages
It is unfortunately implementation dependent as to whether evaluating
DEFPACKAGE forms refering to an existing package modify the existing
package or not. The overloading of the SWANK-LOADER package to contain
both ASDF definitions as well as exporting external symbols breaks the
use of swank.asd under implementations such as ABCL.  We fix this by
delegating the necessary packages to the ASDF DEFSYSTEM loading
mechanism.

Supercedes <https://github.com/slime/slime/pull/537>.