slime-dispatching-connection cleanups
This removes an unnecessary binding and defvar.

slime-connect binds slime-dispatching-connection before calling
slime-setup-connection, which in turn binds this variable to the same
value.  There were previously other functions called during the scope
of this binding in slime-connect, but those calls have since been
removed.

There has been a defvar for slime-dispatching-connection with an
initial value and docstring.  The defvar being removed here was added
later in commit 55ef9ac0 when enabling lexical binding for slime.el,
and this defvar was presumably only for making the binding in
slime-connect dynamic.  With the removal of this binding this defvar
is unnecessary.