Less obvious changes.
First, replace function `second` with `cl-second`.

Second, rename undefined reference `name` to the bound variable `what`.

Lastly, replace `font-lock-fontify-buffer` with `font-lock-ensure`.

The documentation for `font-lock-fontify-buffer` says to use
`font-lock-ensure` or `font-lock-flush` instead.

However, this may not be a good idea---they call functions stored in
different variables, so the behaviour might be unexpected.