abcl: restore conditional guards for JSS
Somehow, some of the conditional guards for the presence of JSS, had
acquired a extraneous character.  We fix that, but note that running
SLIME under ABCL in the absence of loading the JSS contrib does not
fully work as expected.  Since the usual deployment is that JSS will
be available by ACBL-CONTRIB, this should not effect the majority of
usage scenarios.