variable *find-definitions-all-packages*, function find-definitions
Doc: New feature. Suppose you do meta-. and there is no definition for
the symbol in the current package. In that case check for symbols with
the same name in other packages. Only do this if
*find-definitions-all-packages* is t, default nil.

--

Source: eval-region

Doc: New feature. In the repl, having an extra ending extra
parentheses would cause entry to the debugger. That's
overkill. Instead warnd. This is currently unconditional but could be
controlled by customization variable. If there's a different change
that would lead to same behavior I am interested in hearing it.

--

Source: ed-in-emacs

Doc: New feature. Extend ed-in-emacs so a string can be passed to be
put in a buffer. Used, for example, when an inspector has a long
string that you want to play with. Syntax (:string ...)

--

Soure: prepare-part

Doc:
1) New feature. Allow :multiple to collect a set of parts. Used in a
context where a single part is expected, e.g. lcons*

2) New feature. New part type: :strong-value, which can have different
face from value. Would be better to have a general mechanism for
specifying face in inspector parts but that's more work. Alternative
allow :face keyword arg in :value, :label, :action

3) Bugfix. For parts like (:line label value), put the label inside a `(:label ,part).

--

Source: emacs-inspect (array)

Doc: Bugfix. Make the label "Contents: " uses :label.
Sort all numbers, not just reals