add a suffix to the name of inferior lisp buffers
In the current state, SLIME names all inferior lisp buffers `*inferior-lisp*`
by default. When starting multiple implementations, we end up with multiple
inferior lisp buffers whose only difference is a non-descriptive number
suffix.

With this patch, we use the name of the implementation as suffix, in a way
similar to REPL buffers.

For example with two inferior lisps SBCL and CCL, we now have `*inferior-lisp
sbcl*` and `*inferior-lisp ccl*`

We also name comint processes the same way, but with a dash since there are no
"*" characters around to delimit the name.