Rework swank.asd to produce actual compilation artifact (#760)
* Rework swank.asd to produce actual compilation artifact

Until now swank.asd compiled the system files when it was loaded. That is
roughly compatible with what slime does when loading swank, however that is not
useful for swank that is loaded as a standalone system without slime. Instead of
producing compilation artifacts that may be further processed by asdf (either in
bundle-op or otherwise) it puts fasls in ~/.slime directory.

When a system "foo" depends on "swank" then before the system "foo" is loaded
swank is loaded itself. However if we only compile systems and then try to load
a bundle then all we have is the swank-loader artifact - no "SWANK" package nor
nothing. This issue has been unnoticed because usually people dump images
instead of having separate compilation and loading steps.

This commit defines two systems "swank" and "swank-exts" - one for the core
runtime and one for contribs.

* swank.asd: inhibit loading swank only when started from emacs

This inhibition is tailored for possibly incompatible slime/swank
instances (according to the comment at the top of the file), so there is no need
to prevent loading a different swank version in the image without slime.

* Remove locking of the system "swank"

Don't prevent reloading the system "swank" even if it was previously loaded by
emacs. If the user wants to reload swank from a different directory then so be
it. Otherwise we can't compile bundles that depend on slime.

* meta: add an entry to NEWS section