Fix emacs error when M-.-ing to nonexistent file
When the Lisp reports a source location is in a file that doesn't
exist, and also gives the :align t hint, slime-location-offset
will try to (slime-forward-sexp) (beginning-of-sexp) in a new
empty buffer. This causes a pretty inscrutable
"Wrong type argument: characterp, nil" error from emacs. This
problem may also occur if the source file does not have sexp
syntax, but I haven't tried that myself.

The change just silently ignores any errors from the sexp motion.
It might be better to have slime report that the file isn't real
and not do any motion to begin with.