Update ~/.slime to use $XDG_DATA_HOME
This commit updates the `~/.slime` references in `Makefile` and
`swank-loader.lisp` to check for XDG_DATA_HOME. If the environment
variable is set, it uses $XDG_DATA_HOME/slime as the default fasl
directory.

This change is necessary as there is no way to avoid creation of
"~/.slime" when loading a system that requires Swank. My personal
example is trying to port a Clack application to OpenBSD. The OpenBSD
build system does not allow writing to home directories as part
of the build process.

Prior to these changes, I tried using asdf:defsystem-depends-on,
asdf:around-compile, and use of a secondary system to only load
swank.