fix HANDLER-CASE indentation
Indentation of HANDLER-CASE error clauses is currently confusing when the
condition type specifier is a list spanning multiple lines. The most common
case is a OR type specification too long for a single line:

    (handler-case
        nil
      ((or foo-error bar-error
        my-very-long-condition-name)
        (condition)
        nil)
      (baz-error
        (condition)
        nil)
      (error (condition)
        nil)
      (:no-error
        (a b c)
        nil))

The OR form is incorrectly indented, and the second element (the list
containing the condition variable name) is indented at the same column as the
following form(s).

This patch modifies indentation so that the type specification list is
correctly indented. It adds one indentation level to the condition variable
list to make it easier to distinguish from subsequent forms.

    (handler-case
        nil
      ((or foo-error bar-error
           my-very-long-condition-name)
          (condition)
        nil)
      (baz-error
          (condition)
        nil)
      (error (condition)
        nil)
      (:no-error
          (a b c)
        nil))