Signal user-errors instead of errors for user facing interactive functions
Using `error` for signaling errors cause entry to the Emacs debugger
if it is enabled.
For interactive user facing functions it is better to use `user-error`
instead, that signals an incorrect manipulation by the user, not an
actual problem.

Docs of `user-error`:

"This is like error except that a user error (or "pilot error") comes
from an incorrect manipulation by the user, not from an actual problem.
In contrast with other errors, user errors normally do not cause
entry to the debugger, even when debug-on-error is non-nil."