Separate '-autosearch-mode' and '-paredit-compatibility'.
This refactors the functionality so that instead of one custom
variable, 'slime-repl-history-recall-mode', there are now two,
'slime-repl-history-autosearch-mode' and
'slime-repl-history-paredit-compatibility', which can be set
orthogonally.  The former controls only the behavior of M-p when the
input line is nonempty; the latter makes a few tweaks so that things
work nicely for users of ParEdit and similar balanced-parenthesis
packages.