Fix for slime.info path for MELPA package
By default, the function `slime-info' assumes that the location of
slime.info is to be picked up from a sub-directory where the source
is. But that doesn't work if the package is packaged & installed via
MELPA since `slime.info` sits in the top level directory, where `doc'
directory is omitted.

The fix is to let the user know where to look for slime.info to create
it, if they are using slime from source. Or to report it upstream, to
the package manager maintainers, i.e. MELPA.