common-lisp-hyperspec-root path handling
When `common-lisp-hyperspec-root` is set without the leading
`file://`, the call to hyperspec-lookup is dropped silently.

When `common-lisp-hyperspec-root` is set without the trailing `/` (on
Unix like OS), the hyperspec-lookup is also dropped silently.

Ideal fix for this would be to use something like this:

(defcustom common-lisp-hyperspec-root
  "http://www.lispworks.com/reference/HyperSpec/"
  "The root of the Common Lisp HyperSpec URL.
If you copy the HyperSpec to your local system, set this variable to
something like \"/usr/local/doc/HyperSpec/\" and it will automatically
be prefixed with \"file://\" and normalized as a proper directory path."
  :type 'string
  :set (lambda (symbol value)
         (let ((processed-value
                (if (and (stringp value)
                         (not (string-match-p "^https?://" value))
                         (not (string-match-p "^file://" value)))
                    (concat "file://"
                            (file-name-as-directory
                             (expand-file-name value)))
                  value)))
                  (set-default symbol processed-value))))

But since SLIME is over 23 years old, and hyperspec.el is over 28
years old, there might be some backwards compatibiliy issues.

Because you can't set defcustom values with setq.

So instead, this fix tries to treat those edge case potential errors
by checking the value of `common-lisp-hyperspec-root` and ensures that
it has proper formatting, so that the calls aren't dropped silently.